# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Endoquiz API",
        "description": "Rotas do Endoquiz",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://endoquizapi.endoquiz.com.br/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/get_classified": {
            "post": {
                "tags": [
                    "Endoquiz"
                ],
                "summary": "GETALLCLASSIFIED",
                "description": "Rota para retornar todos os classificados com os tópicos",
                "operationId": "571a4e3fbba3261bd3b235b05795b41f",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Informe o token do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idUser",
                        "in": "query",
                        "description": "Informe o id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/get_classified_subtopic": {
            "post": {
                "tags": [
                    "Endoquiz"
                ],
                "summary": "GETALLCLASSIFIEDSUBTOPICS",
                "description": "Rota para retornar todos os subtópicos e assuntos de algum topico do classificado",
                "operationId": "25db67d1d6a659b13b5bb337e355f319",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Informe o token do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idUser",
                        "in": "query",
                        "description": "Informe o id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idTopic",
                        "in": "query",
                        "description": "Informe o id do tópico",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}